requirements.txt
appium-python-client>=3.1.1
pytest>=7.0.0
allure-pytest>=2.13.5
selenium>=4.10.0
